fun spear

+++ src/game_engine/world.rs
@@ -707,7 +707,7 @@ impl World{
                     }
                     let length = attack.stats.size.unwrap_or(0.0).floor() as usize;
                     let width = (attack.width_to_length_ratio * length as f32) as usize;
-                    let collisions = self.get_attacked_rotated_rect(true, None, attack.x as usize - length/2, attack.y as usize - width/2, length, width,attack.angle, true);
+                    let collisions = self.get_attacked_rotated_rect(true, None, (attack.x - length as f32/2.0) as usize, (attack.y - width as f32 /2.0) as usize, length, width,attack.angle, true);
                     let mut hit = false;
                     for collision in collisions.iter(){
                         if self.entity_health_components.contains_key(collision){
@@ -721,7 +721,7 @@ impl World{
                         let aoesize = attack.stats.AOE.unwrap_or(0.0);
                     let length = (attack.stats.size.unwrap_or(0.0).floor() + aoesize) as usize;
                     let width = (attack.width_to_length_ratio * length as f32 + aoesize) as usize;
-                        let aoe_collisions = self.get_attacked_rotated_rect(true, None, attack.x as usize - length/2, attack.y as usize - width/2, length, width,attack.angle, true);
+                        let aoe_collisions = self.get_attacked_rotated_rect(true, None, (attack.x - length as f32/2.0) as usize, (attack.y - width as f32 /2.0) as usize, length, width,attack.angle, true);
                         for collision in aoe_collisions.iter(){
                             if self.entity_health_components.contains_key(collision){
                                 let health_component = self.entity_health_components.get(collision).unwrap().borrow_mut();
